Simplify the action for enabling/disabling a user's avatar/signature/cover photo.
