+++ wcfsetup/install/files/lib/action/UserDisableSignatureAction.class.php
@@ -2,14 +2,7 @@
 
 namespace wcf\action;
 
-use Laminas\Diactoros\Response\JsonResponse;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Server\RequestHandlerInterface;
-use wcf\data\user\group\UserGroup;
 use wcf\data\user\UserProfile;
-use wcf\http\Helper;
-use wcf\system\cache\runtime\UserProfileRuntimeCache;
 use wcf\system\exception\IllegalLinkException;
 use wcf\system\exception\PermissionDeniedException;
 use wcf\system\form\builder\field\BooleanFormField;
@@ -28,64 +21,27 @@ use wcf\system\WCF;
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since 6.3
  */
-final class UserDisableSignatureAction implements RequestHandlerInterface
+final class UserDisableSignatureAction extends UserManagementAction
 {
     #[\Override]
-    public function handle(ServerRequestInterface $request): ResponseInterface
+    protected function performAction(UserProfile $userProfile, array $data): void
     {
-        $parameters = Helper::mapQueryParameters(
-            $request->getQueryParams(),
-            <<<'EOT'
-                array {
-                    id: positive-int
-                }
-                EOT
-        );
-
-        $user = UserProfileRuntimeCache::getInstance()->getObject($parameters['id']);
-        $this->assertSignatureCanBeDisabled($user);
-
-        $form = $this->getForm();
-
-        if ($request->getMethod() === 'GET') {
-            return $form->toResponse();
-        } elseif ($request->getMethod() === 'POST') {
-            $response = $form->validateRequest($request);
-            if ($response !== null) {
-                return $response;
-            }
-
-            $data = $form->getData()['data'];
-            $reason = $data['reason'];
-            if ($data['neverExpires']) {
-                $expires = null;
-            } else {
-                $expires = $data['expires'];
-            }
-
-            (new DisableSignature($user->getDecoratedObject(), $reason, $expires))();
-
-            return new JsonResponse([]);
+        $reason = $data['reason'];
+        if ($data['neverExpires']) {
+            $expires = null;
         } else {
-            throw new \LogicException('Unreachable');
+            $expires = $data['expires'];
         }
+
+        (new DisableSignature($userProfile->getDecoratedObject(), $reason, $expires))();
     }
 
-    private function assertSignatureCanBeDisabled(?UserProfile $userProfile): void
+    #[\Override]
+    protected function assertUserCanBeManaged(?UserProfile $userProfile): void
     {
-        if (!$userProfile) {
-            throw new IllegalLinkException();
-        }
-
-        if ($userProfile->userID === WCF::getUser()->userID) {
-            throw new IllegalLinkException();
-        }
+        parent::assertUserCanBeManaged($userProfile);
 
         if (!WCF::getSession()->getPermission('admin.user.canDisableSignature')) {
-            throw new PermissionDeniedException();
-        }
-
-        if (!UserGroup::isAccessibleGroup($userProfile->getGroupIDs())) {
             throw new PermissionDeniedException();
         }
 
@@ -94,7 +50,8 @@ final class UserDisableSignatureAction implements RequestHandlerInterface
         }
     }
 
-    private function getForm(): Psr15DialogForm
+    #[\Override]
+    protected function getForm(): Psr15DialogForm
     {
         $form = new Psr15DialogForm(
             static::class,
